Update navigations for review detail and privacy settings
